requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.22
streamlit==1.29.0
python-dotenv==1.0.0
pypdf==3.17.4
python-docx==1.1.0
